public/private api
