+++ src/index.ts
@@ -1,22 +1,5 @@
 export * from "./types"
-export {
-    loginByPasswordFx,
-    loginByTokenFx,
-    initStoreFx,
-    startClientFx,
-    stopClientFx,
-    searchRoomMessagesFx,
-    sendMessageFx,
-    editMessageFx,
-    deleteMessageFx,
-    readAllMessagesFx,
-    getRoomsWithActivitiesFx,
-    getRoomInfoFx,
-    getLoggedUserFx,
-    initRoomFx,
-} from "./effects"
-export * from "./events"
-export * from "./computed"
+export * from "./public"
 export * from "./matrix-client"
 export * from "./utils"
 import "./init"
